// src/app/core/model/clientes.model.ts
export class Cliente {
    id: number;
    name: string;
    federalId: string;
    registration: string;
    phone: string;
    phone2: string;
    email: string;
    emailCollection: string;
    residentialPhone: string;
    commercialPhone: string;
    emergencyContact: string;
    emergencyPhone: string;
    birthday: Date;
    gender: string;
    federalIdType: string;
    commercialAddress: string;
    residentialAdress: string;
    active: string;
}